requirements.txt
pyautogui>=0.9.54
Pillow>=10.0.0
easyocr>=1.7.0
openai>=1.0.0
pywin32>=306
pyyaml>=6.0
pyperclip>=1.8.2
keyboard>=0.13.5
pyvda>=0.3.0
fastapi>=0.110.0
uvicorn>=0.27.0
websockets>=12.0
python-multipart>=0.0.9
